pre/SparseCT/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astra-toolbox=1.9.9.dev6=py38h452e1ab_0
blas=1.0=mkl
ca-certificates=2021.1.19=haa95532_1
certifi=2020.12.5=py38haa95532_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
git=2.23.0=h6bb4b03_0
icc_rt=2019.0.0=h0cc432a_1
imageio=2.9.0=pypi_0
intel-openmp=2020.2=254
kiwisolver=1.3.1=pypi_0
libastra=1.9.9.dev6=hfa6e2cd_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
networkx=2.5=pypi_0
numpy=1.20.1=pypi_0
numpy-base=1.19.2=py38ha3acd2a_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1j=h2bbff1b_0
pillow=8.1.2=pypi_0
pip=21.0.1=py38haa95532_0
pydicom=2.1.2=pypi_0
pyellipsoid=0.0.3=pypi_0
pyparsing=2.4.7=pypi_0
python=3.8.8=hdbf39b2_4
python-dateutil=2.8.1=pypi_0
pywavelets=1.1.1=pypi_0
scikit-image=0.18.1=pypi_0
scipy=1.6.1=py38h14eb087_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
sqlite=3.35.2=h2bbff1b_0
tifffile=2021.3.17=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
